+++ homeworks/brisa-rojas/21-csr-ssr-ssg-nextjs/src/components/AvatarTile/AvatarTile.js
